feat[Block Broadcast]: Add block broadcast to the network
Add block broadcasting. The validator which is statically selected right now
is able to create a new block every block time and broadcast it to the network.
The other non validator nodes will receive the block and validate it, and if
valid, add it to their blockchain, and then broadcast it to the network, resulting
in gossip behavior.

Note: Add a late / out of sync server in main.go to test the blockchain synchronization.
This currently fails and is not implemented yet.

All unit tests pass.