feat[Synchronization]: Add synchronization between peers
Add synchronization for late joining nodes, or nodes that have failed
and are rejoining the network.
Limitation: There is a limitation currently with the buffer size
we read from the TCP socket. This will be fixed in a future commit.
We will need to download the peer blockchain in parts instead of sending
the entire thing over the network in one go.

All unit tests pass.