feat[Block Reward]: Add reward coinbase transaction for validators that mine a block
Add block reward coinbase transaction when a new block is created.
The transaction is from the coinbase to the account of the validator node.
This ensures that there is an incentive for nodes to mine blocks honestly.

TODO: need to change how the coinbase transactions are made to be more secure.

All unit tests pass.